ready to update atss metrics

+++ utils/model_info.py
@@ -3,11 +3,12 @@ from copy import deepcopy
 import torch
 import torch.nn as nn
 from thop import profile
+from torch.nn import parameter
 
 def get_model_info(model, tsize):
 
     stride = 64
-    img = torch.zeros((1, 3, stride, stride), device=next(model.parameters()).device)
+    img = torch.zeros((1, 3, stride, stride), device=next(model.parameters()).device, dtype=torch.double)
     flops, params = profile(deepcopy(model), inputs=(img,), verbose=False)
     params /= 1e6
     flops /= 1e9
